Support invitations from anonymous support users
This changes the invitations controller so that if the current inviter
is an anonymous support user, this isn't used to create the invite. This
is because the anonymous user isn't in the database and when setting the
foreign key on the user, it fails.

See https://github.com/DFE-Digital/apply-for-qualified-teacher-status/pull/166/commits/688120f1855cc484da5cbe6d606ca43b4abb1372