Initialise a Rails engine for the feature flags
We want to isolate the feature flags code from the rest of the app.

The intention is that this is a common piece of functionality across
projects in DfE Digital. It would be useful to extract it into a
separate library to make it easier to share the implementation.

A first step towards this is to isolate the code using a Rails Engine.
This would make it easier to extract to a gem in the future.