Add configuration for feature flags
We want to provide a way to version the addition of feature flags while
also ensuring the value for the flag is dynamic and persisted in the
database.

This is based on the [implementation in
Find](https://github.com/DFE-Digital/find-a-lost-trn/pull/57).

A YAML configuration file has been introduced to accommodate the Rails
Engine approach. We need a way to configure the list of available flags
from the parent app, without needing to modify the engine code.

The UI for managing the state of the flags will follow in a subsequent
commit.