Provide a UI for managing feature flags
The configuration of the feature flags was setup in the previous PRs but
they didn't provide a UI to modify the dynamic values of the flags.

This change introudces a dashboard to manage the feature flag values.

It is based on [the one used in Find](https://github.com/DFE-Digital/find-a-lost-trn/pull/76).