Fix bundling the engine assets
When providing assets via an engine, they don't get included in the
esbuild command.

This change ensures that the engine defines its own step for building
CSS and JS via esbuild.

Ensuring that the output files are namespaced in the app/assets/builds
folder is enough to allow the parent app access to them.