Add the reporting as question
The first question of the eligibility screener asks for the context of
the report you are making.

`EligibilityCheck` is the representation of the outcome of the screener.

As per the approach in Find a Lost TRN and Access Your Teaching Profile,
we're utilising form objects to represent a question and persisting the
answers to the `EligibilityCheck` entity.

There is no session persistence yet but this will be added when we have
multiple questions.