Add a page with links to the relevant referral form
When someone completes the eligibility screener, they should see
guidance on the appropriate referral form they need to use.

Currently, the forms are hosted on GOV.UK so we opted to hotlink to them
and accept the maintenance that comes with hotlinking. This is
preferable over having to host them in the service and keep them in sync
with the GOV.UK ones.

This change introduces a simple session based method to persist teh
eligibility check across multiple steps of the form.

It is anticipated that this will evolve into something like the
solutions on Find and Access.