Add the serious misconduct screener question
One of the questions in the screener asks if you are reporting serious
misconduct.

Choosing "no" prompts you to report the issue somewhere else.

Both "yes" and "I'm not sure" take you to the next step in the flow.

The "I'm not sure" option also displays a hint as to what to do next.

This question was created using the recently added form generator.

There is no back link present, nor any object orchestrating the question
order. These will be added in subsequent PRs.

I was using this question as a qay of verifying the form generator was
configured appropriately for the refer project.