Add a support interface for eligibility checks
We want a page in the support interface to provide some visibility into
the eligibility checks that are created by this service.

Taking inspiration from [Find](https://github.com/DFE-Digital/find-a-lost-trn/pull/53), this change introduces
a basic implementation of a list of `EligibilityCheck`s.

The list is limited to the most recent 100 records. When the list
starts to exceed that limit in production, we can look at adding
pagination.

Also at this point, we haven't added any data anonymisation to this page
as there is nothing personal or specific stored yet. When that changes
we can revisit this decision.