Enforce question order
We want to extract the question order into a single location so that
changes to the order don't cascade throughout the form objects.

This is based on the implementation in Find.

`EnforceQuestionOrder` holds all the logic required to derive the next
question for the current eligibility check.

The last couple of screens in this flow are exceptions to the
implementation as they aren't questions with a persistable answer.