Add the "is a teacher" question
We want to determine if the allegation is about a teacher.

This question introduces a new fork and exit page for the flow.

While doing this the complexity of the enforce question order class was
growing because it was handling the non-answerable pages.

To simplify, I moved the logic for these exit pages to the controller
that precedes them. This allows the question ordering class to return to
a single responsibility.