Change how EnforceQuestionOrder is included in controllers
EnforceQuestionOrder is currently included in the top-level
ApplicationController. We'll shortly be introducing a number of
controllers that don't require the logic in EnforceQuestionOrder and
shouldn't be subject to the before_action redirect it provides. Rather
than manually skip the before_action wherever it isn't needed, instead
do the following:

- Remove EnforceQuestionOrder from ApplicationController
- Add it explicitly in all controllers related to the eligibility
  checker

Incidentally, this allows us to restore the inheritance relationship
between SupportInterfaceController and ApplicationController.