Add static pages written in markdown
The static pages don't contain any custom HTML and are simple text
documents.

It makes most sense to write them in markdown to allow for ease of
reading and writing in development.

The markdown rendering implementation is similar to the one implemented
in [Find a Lost TRN](https://github.com/DFE-Digital/find-a-lost-trn/pull/21).