Replace debug with pry and pry-nav
The debug gem is fine for most use cases, but currently doesn't work
well with Cuprite, the Capybara driver we're using. After a
`binding.break`, the debug console loses its connection to the browser.
This means commands like `save_and_open_page` or `save_screenshot` don't
work.

Until this is fixed, we can use pry. We add pry-nav to give us the
'continue', 'step', and 'next' commands in the debugger.

Example of the Cuprite bug:
https://github.com/openfoodfoundation/openfoodnetwork/pull/9533#issuecomment-1210618186

An open issue on the debug gem highlighting the problem:
https://github.com/ruby/debug/issues/295