Collect the referrer's name
The first piece of information we collect on the referrer is their name.

An assumption here is that simply the presence of text is enough to
proceed with the referral.

In the designs, the referrer details screen has a way to indicate
whether the flow has been completed or not. This is left to a future PR
to avoid this one getting overly complex.