Add Contact details/Check answers page
At the end of the contact details journey a page with an overview of
the teacher's contact details they submitted. A "Change" link allows
them to go back to that page and change the answers.

The section can be marked as complete or as still in progress. An error
message will be shown when trying to submit without selecting "Yes" or
"No".

Use factories for referral