Add check answers for the organisation details
When a user adds their organisation details, we display a summary of the
information on a check answers page.

They also have the ability to indicate if the details are complete or
not.

I opted to use a separate model, `Organisation`, to store the organisation
details rather than overloading `Referral`.