Simplify the smoke test
We have a user authentication flow behind a feature flag. When the
`user_accounts` flag is switched on, this authenticates everything
beyond the start page.

This causes some issues with the current smoke test. It's written to
exercise the eligibility screener without any user authentication. If
the `user_accounts` flag is active (in one of the deployed test
environments, for example), the smoke test will fail. We could
deactivate the feature via the smoke test, but all it would take to halt
a deploy is for someone to activate it again manually via the Support
UI.

The smoke test currently covers the entire eligibility screener. Both
this and the user sign in process are extensively covered in our system
specs. As such, this change reduces the scope of the smoke test to
simply check the following pages load successfully:

- the start page
- the first page of the screener