Removed terraform prevent_destroy on storage
This setting provides protection for the accidental deletion of the
storage resource by terraform.  They don't prevent deletion by other
services or users.  It isn't possible to deploy a review storage account
with this setting as it can't be applied dynamically.  This change
removes prevent_destroy to allow for review app deployment.