Control access to the screener using a feature flag
We want to control access to the eligibility screener and redirect to a
more appropriate location in the case where we need to disable it.

This allows the GOVUK page to direct people to the screener and we can
control whether we present the new flow for people or simply redirect
them to the existing location on GOVUK.

The main benefit here is that we don't need to ask another team to
modify the redirect in the case of issues with the screener.

I considered using a `before_action` in a base controller to implement
the redirect, but then thought it might add a layer of indirection for
the maintainer of the code.

Using a constraint to restrict access to the screener routes in the
route configuration seemed the simplest and clearest for someone reading
the code.

Assumptions:
* the redirect is temporary, so I used a 307 status code
* the feature flag is temporary, so having it in a single file makes it
  easier to remove in the future
* the redirect location on GOVUK is static, so hardcoding it is fine vs
  an ENV