Skip the smoke test temporarily
While we're rolling out the eligibility screener, the feature flag
arrangement breaks the smoke test.

We're unable to run the smoke test successfully while the feature flags
are disabled in production.

This prevents us from deploying automatically.

While we roll out the screener and watch for issues, we're disabling the
smoke test.

The intention is that this will only be for a few days.