Collect a detailed account of previous allegations
When someone knows of previous misconduct, we want to allow them to
provide a detailed account of it as part of the referral.

This change adds a text field for someone to provide details of previous
misconduct. Uploading a file that documents the previous misconduct will
happen in a follow on change.