Add job duties page
The teacher's duties page follows the same organisation page in the
"About their role" journey. This commit also refactors the system spec
by categorising the methods. When editing the job duties page it shows
the current uploaded file, deletes the previous file when uploading
again and allows to go through the journey and not update your answers.