Display the confirmation page
After submitting the referral, we want to display a confirmation page to
the referrer with instructions about what to expect to happen next.

In order to be able to submit the referral, all sections need to be
complete.

There is some difference in the values of the status fields for each
section. The validation on the form is expecting it to be `complete` but
the sections themselves use either `complete` or `completed`.

The correct styling of the label is triggered by the value `completed`,
which is also the word used in the designs.

This change ensures we use a consistent status value for all sections.

I made a decision to use the `edit` action to represent the summary page
in edit mode and the `show` action will be the review page. Currently,
they both render the same view.