Add a lint check for RSpec
There are variations in how we write our specs, which will likely
continue to vary as other people contribute to the project.

To reduce the cognitive load on the author when writing and the
maintainer when reading, we can auto-format our specs using the RSpec
cops for Rubocop.

These cops mostly follow the style described in `https://betterspecs.org`.

The exceptions added to `rubocop.yml` are to ensure the cops match the
testing styleguide we already have written.

Feel free to modify the cops if you have a strong opinion on any of
them.

This change brings in the default cops and ensures our codebase doesn't
have any violations.

Also, while there I noticed we have 2 time helpers included in the
project. Seeing as Rails provides this already, it makes sense to ditch
Timecop.