Add Teacher role / Check answers page
At the end of the teacher role journey display a page with an overview
of the details they submitted. A "Change" link allows them to go back to
that page and change the answers.

The section can be marked as complete or as still in progress. An error
message will be shown when trying to submit without selecting "Yes" or
"No".

This commit also adds the redirect back to the check answers page if the
page was accessed directly by clicking the "Change" button.