Add a declaration page
In order for someone to submit a referral, they need to agree to a
declaration statement.

This change introduces the new screen and a timestamp field to represent
when the agreement occurred.

I opted for the timestamp due to the extra dimension of data it
provides, which is particularly useful for calculating how long it took
to create a referral, from creation to submission.