Add test users to support interface
Introduce functionality that allows support staff to sign in with a test user
account.

The Test Users section of the support interface lists user accounts that
have been created on the environment. Support staff can either sign in
as an existing user or create a new one.

This functionality is restricted to test environments.