Make the back link aware of the return to value
The back link on each page is a simple `url_for(:back)` but this doesn't
recognise a user visiting from a check answers page.

We already store a return to location in the session to allow redirects
back to a check answers page when appropriate.

The simplest change to allow the back link to behave in a similar
fashion is to use this same session value.

Whenever someone clicks a change link from a check answers page, both
the back link and submitting the form successfully should take the user
back to the check answers page.

I considered taking an approach similar to the one in
[Apply](https://github.com/DFE-Digital/apply-for-qualified-teacher-status/pull/787)
but decided that the complexity and changes required outweigh the
benefits for this project.

Instead, I opted for the simpler (and less feature rich) approach of
using the value in `session[:return_to]` and falling back to a
hard-coded default in the view.

The upside of this is that the URLs are defined explicitly, and at an
individual page level it is easy to understand the navigation.

The downside is that every page needs to define it's referrer.

I think the trade-off is worth it.