Add eligibility_check foreign key on referrals
We want to associate the eligibility check database record with the
referral that gets created after the check is complete. This will allow
us to determine whether the referral is submitted by an employer or a
member of the general public.

As a first step, add the foreign key column to the referrals table.