Swap factorybot usage for User model when creating test users
We currently use factorybot when creating a new test user. We don't have
the factorybot gem available outside of dev or test groups so this fails
on deployed environments. Rather than move it into the ungrouped set of
gems for this one use case, this change swaps FactoryBot for direct
interaction with the User model.