Add data migration to remove Referrals without eligibility checks
We've recently reworked the flow through the system such that completing
the eligibility screener results in the creation of the referral, with
an association set up between the two. The association is mandatory by
virtue of the belongs_to that sits on the Referral model.

A side effect of this change is that existing referrals in the various
environments will trigger a validation error when we attempt any kind of
save.

As we haven't gone live yet, a simple way to fix this is to delete old
referrals without associated eligibility check records.