Reduced resource requirements for ACI worker app
The Azure Container Instance that runs the sidekiq worker app typically
uses 200MB RAM and 2% CPU (bursting to 200 ~ 1 p/day).  The higher the
resource demands the more likely that the Azure region we are deploying
into won't be able to satisfy that requirement.  Reducing resource
demand increases reliability of deployment