Remove job title from the public flow
When someone is referring as a member of the public, we don't need to
collect their job title. It is not relevant.

I considered creating a new route and controller for this flow, as per
the approach in #297.

This seemed overkill as the only difference we have in this particular
case is to change the value of `next_path` for the name controller.

It seems like we're missing an orchestration object where we could
define the questions and their order in a more declarative fashion.

I opted not to do that here as it doesn't seem like the right time to do
it.

Instead, I went for the simplest change possible, which is to use a
conditional statement to determine the next question in the name
controller.