Add a check for referral type in controllers
When retrieving a referral from the database, it would be useful to have
a check which ensures that the referral being used in any given
controller is the correct type for the current namespace (ie - public vs
employer referral).

This commit adds a controller concern which raises an error if the
current path doesn't correspond to the referral being retrieved.

This ensures we don't accidentally pass users into the employer flow
while building out the public one.