Replace subsection_path with polymorphic_path
Rails provides a helper method for generating URLs that our
`subsection_path` method duplicates.

We can replace the custom one with the Rails one to simplify
understanding for future maintainers.

The critical addition needed to allow us to use the Rails helper is a
reliable way of determining the routing scope.

I added a new method to `Referral` for this purpose, `#routing_scope`.

Ideally, I would have liked to use the value of
`EligibilityCheck#reporting_as`, as it returns either `employer` or
`public` but because we have only implemented routes specific to
`public` and left the employer routes as implied, I had to create a new
method.

If we continue to place the scope in the URLs then I think for consistency
purposes we implement the employer routes in an `employer` namespace as
well.