Ensure the check answers pages have back links
Not all the check answers pages have a back link and the ones that do
often link to the previous page.

The latest designs call for the back link on the check answers page to
always return to the main page for the referral.

While making this change, I noticed that the 'Your details' back links
only work in the context of an employer referral. There is a fix for
that here too.