Add a link to the uploaded allegation
When a user uploads a file as their allegation, we don't currently link
to the file.

This is inconsistent with other uploaded files in the app.

Of note in this change is that I moved the location of the
`allegation_details` method from the associated form object to the view
component.

This is to ensure access to the view context so that the link to the
filename can be rendered easily.