Use scoped routes where possible
Now that the referrals are scoped by type, we need to ensure links
take the scope into account.

The easiest way to do this is to use the Rails link helpers and pass the
routing scope. If it is `nil` then the default scope is used, eg.
`referrals`, alternatively we use `public_referrals`.

There are some views and controllers that are only used for the employer
flow, so I left some of those links as they are.

I attempted to change the uploaded evidence views but ran into an issue
with the routes not matching the model they refer to, eg.
`ReferralEvidence` is rendered at the `referrals/1/evidences/1` route.

The custom route means we need to override more things and I was trying
to limit the scope of this change to simply replacing the helper methods
with the array style.