Add all sections to case worker referral view
This adds all the sections to the case worker referral view that are
detailed in the current designs.

There are some parts to this that aren't implemented yet.

The referrer name is stored as a single field. For now I opted to split
this value into first and last name methods so we can render the
proposed UI.

When the change to the way we collect this value happens, then these
methods will continue to work and we won't need to change the referral
view.

Secondly, we don't collect the referrer's current employer address.

Rather than guess at how we might do that, I left it unimplemented.