Update the name for linked files
The format of the linked files contains a 'File:' prefix.

Since we implemented this, the designs have changed and no longer
require this prefix. We can safely remove it.

While I was making a change there, I took the opportunity to standardise
how the allegation description is formatted.

Both the current allegation description and the previous allegation
description should be displayed in the same way.