Try grover
Rendering a PDF from the controller but the CSS isn't working in this
commit. Grover has a pre-processor that should fix that.

Rendering from the controller lik this is going to mean the ZIP file
request has to make another request back for a different format before
adding the file. It'll be tricky to do the PDF async as there's no user
context.

The existing template won't render without a request as it needs csrf
and other stuff that we'd need to stub out.