Remove sign-in/out related flash messages
We no longer want to display the flash messages around authentication.

We initially wrote specific logic to ensure they are displayed. Rather
than attempt to overwrite Devise to prevent them being generated, I
opted to simply clear the values of the messages.

This gives us the flexibility to use the other parts of the Devise flash
messages while removing the specific sign-in/out ones.