[1279] Remove start page
* Remove the existing start page
* Refactor the start path logic to allow it to be tested more easily

As the 'first page' logic is quite complicated, extract it into a
service class in lib. The existing logic in `ApplicationController` is
also used in the referral create action. We should refactor that too but
it's out of scope of this PR.

Tests added in a separate commit as there were a lot of changes so it's
noisy.