Hide "Not known" rows when the previous answer is "No"
They were showing up as "Not known" or "Not answered" when the previous
answer was "No". This change makes the referral forms consistent with
the manage interface. I also updated the "Their name" row because it
needs to be `nil` to show a "Not answered" row. If the name was missing
the interpolation would return an empty string.