Use CustomAttrs for eligibility pages
Updates the eligibility pages to use custom attributes. This will also
help with showing the selected answers when navigating back to the page
and automatically converting true/false to boolean values.

It also fixes a bug where calling `process_type` for a db string value
that get processed as a symbol. E.g. the `reporting_as` values of
:public, :employer