Refactor file upload to use an Upload model
This commit refactors the file upload functionality to use the Upload
model. This is a polymorphic model that can be attached to any
other model. It is used to store the file data and metadata.

I also added a rake task to transfer the existing file data to the
new model.