Disambiguate filenames in zip file
Add an incrementing index to the file names as we're creating the
downloadable zip file.

This handles the case where the user has uploaded identically named
files to the same section. We could probably handle this more gracefully
as this implementation adds an index to every file name when it is very
rarely required. That could be a further iteration but this fixes a
production issue.